Create real Page and Element wrappers
And add tests for them. Implement just a few methods for now.
