+++ PipeableTests/ScriptTests.swift
@@ -5,7 +5,7 @@ final class ScriptTests: XCTestCase {
     func test_goto() async throws {
         let result = try await runScript("""
             print('Calling page.goto');
-            await page.goto("https://google.com");
+            await fakePage.goto("https://google.com");
             return true;
         """)
 
@@ -15,7 +15,7 @@ final class ScriptTests: XCTestCase {
     func test_goto_with_error() async throws {
         let script = """
             print('Calling page.goto');
-            await page.goto("");
+            await fakePage.goto("");
             return true;
         """
         do {
@@ -30,19 +30,19 @@ final class ScriptTests: XCTestCase {
 
     func test_querySelector_and_click() async throws {
         let script = """
-            const element = await page.querySelector("asd");
+            const element = await fakePage.querySelector("asd");
             await element.click();
             return element;
         """
         let result = try await runScript(script)
         // swiftlint:disable:next force_cast
-        let element = result.toObjectOf(PipeableElementWrapper.self) as! PipeableElementWrapper
+        let element = result.toObjectOf(FakeElementWrapper.self) as! FakeElementWrapper
         XCTAssertEqual(element.successFullClicksClount, 1)
     }
 
     func test_querySelector_and_click_with_error() async throws {
         let script = """
-            const element = await page.querySelector("empty");
+            const element = await fakePage.querySelector("empty");
             await element.click();
             return element;
         """
